server/requirements.txt
# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.0

# Data Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Task Scheduling
apscheduler>=3.10.0

# Logging
loguru>=0.7.0

# Data Processing
# pandas>=2.1.0
# numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Linting
flake8>=6.1.0

